refactor(automations): move run metric/state filter modules up to wandb/automations
Relocate run_metrics.py and run_states.py out of the _filters subpackage
as _run_metric_filters.py and _run_state_filters.py, since they define
automation-specific filters rather than generic filter machinery.
