+++ wandb/automations/events.py
@@ -10,14 +10,14 @@ from wandb._pydantic import GQLBase, model_validator, pydantic_isinstance
 from wandb._strutils import nameof
 
 from ._filters import And, FilterableField, MongoLikeFilter
-from ._filters.run_metrics import (
+from ._generated import FilterEventFields
+from ._run_metric_filters import (
     MetricChangeFilter,
     MetricThresholdFilter,
     MetricVal,
     MetricZScoreFilter,
 )
-from ._filters.run_states import StateFilter, StateOperand
-from ._generated import FilterEventFields
+from ._run_state_filters import StateFilter, StateOperand
 from ._validators import (
     JsonEncoded,
     LenientStrEnum,
